requirements.txt
pyudev
RPi.GPIO